ref: rename file name & remove unused code
