+++ src/api/axios.ts
@@ -6,11 +6,9 @@ export const axiosUser = axios.create({
 
 export const axiosCustomer = axios.create({
   baseURL: "http://localhost:8081/api/customers",
-  // withCredentials: true,
 });
 export const axiosShoppingCart = axios.create({
   baseURL: "http://localhost:8081/api/shopping-cart",
-  // withCredentials: true,
 });
 
 export const axiosProducts = axios.create({
