--- src/components/CartIcon/CartIcon.tsx
@@ -1,7 +1,0 @@
-import { useCartItems } from "../../services/querries";
-
-export const CartIcon = () => {
-  const { data } = useCartItems();
-  console.log("cartItems: ,", data);
-  return <div className=""></div>;
-};
